requirements.txt
Django
django-simple-history
django-unfold
docx2pdf
pillow
python-docx
python-dotenv
gunicorn
psycopg2-binary
